requirements.txt
# Google Cloud Functions framework
functions-framework==3.5.0

# Google Generative AI for Gemini (Google AI Studio)
google-generativeai==0.8.3

# Dialogflow for conversation management
google-cloud-dialogflow==2.26.0

# Additional dependencies for Cloud Functions
flask==3.0.0
